copyrights + cleanup
