+++ components/pages/discovery/charts/Doughnut.tsx
@@ -1,37 +1,29 @@
-import { UikitTheme, useTheme } from '@icgc-argo/uikit';
-import Charts from 'charts';
+/*
+ * Copyright (c) 2025 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ * This program and the accompanying materials are made available under the terms of
+ * the GNU Affero General Public License v3.0. You should have received a copy of the
+ * GNU Affero General Public License along with this program.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import { useTheme } from '@icgc-argo/uikit';
+import Doughnut from 'charts/Doughnut/ui';
 import { Chart } from './Chart';
+import { chartThemeFn } from './theme';
 import { injectTheme } from './util';
 
-const chartThemeFn = (injectedTheme: UikitTheme) => ({
-  theme: {
-    text: {
-      fontFamily: 'Work Sans,sans-serif',
-    },
-    axis: {
-      legend: {
-        text: { fontSize: 10, color: injectedTheme.colors.grey },
-      },
-      ticks: {
-        text: {
-          fontSize: 11,
-          color: 'black',
-        },
-        line: {
-          strokeWidth: 0,
-        },
-      },
-      domain: {
-        line: {
-          stroke: injectedTheme.colors.grey_2,
-          strokeWidth: 1,
-        },
-      },
-    },
-  },
-});
-
-const doughnutChartData = [
+const TEMP_DATA = [
   {
     id: 'Biliary Tract',
     label: 'Biliary Tract',
@@ -90,11 +82,15 @@ const DoughnutChart = ({ field }) => {
 
     colorBy: 'indexValue',
     theme: chartTheme.theme,
+
+    // temp data
+    data: TEMP_DATA,
   };
 
   return (
     <Chart>
-      <Charts.Doughnut field={field} consumerConfig={config} />
+      {/* <Charts.Doughnut field={field} consumerConfig={config} /> */}
+      <Doughnut config={config} data={TEMP_DATA} />
     </Chart>
   );
 };
